Silence javascript alert pop-up
Don’t alert us with a pop-up when there’s a problem looking up an
address. It’s likely because of a badly formatted address or weird
symbols, which we can’t really do much about. Instead report it in the
console so only devs are likely to see it.